Split water_treatment.rs into sub-modules (972 -> 5 files, max 354 lines)
Split the 972-line water_treatment.rs into a directory module:
- mod.rs (19 lines): module declarations and re-exports
- types.rs (249 lines): TreatmentLevel enum, PlantState, WaterTreatmentState, helper fns
- systems.rs (151 lines): update_water_treatment system and WaterTreatmentPlugin
- tests.rs (354 lines): type and effluent quality unit tests
- tests_risk_demand.rs (240 lines): disease risk, demand, and integration tests

All public API re-exported from mod.rs; no changes to lib.rs or external consumers.
